HomeWork/Ky 1/(PPR501.6)Python/Group Project/youtube-link-transformation-be-main/requirements.txt
fastapi==0.100.0
python-dotenv==1.0.0
openai==0.28.0
youtube-transcript-api==0.6.0
uvicorn==0.22.0